mcp/mcp-server/requirements.txt
google-api-python-client==2.122.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
mcp  # Assuming mcp is available on PyPI or in local dev
